requirements.txt
plaid-python==9.8.0
python-dotenv==0.21.0
pandas==2.0.3
schedule==1.2.0